fix issues with recent MooseX::Traits::Pluggable
We were using it in a rather stupid way, applying it to our adapter
then telling it to create a completely unrelated class. Recent
versions (0.11 and later) break with this usage. Now, we subclass the
adapted class adding MooseX::Traits::Pluggable to it, which makes the
whole thing work wuch better, and without relying on undocumented
internals.